requirements.txt
yt-dlp
openai-whisper
youtube-transcript-api
google-generativeai
python-dotenv
torch
setuptools-rust
